fix(statistics): keep the final Geyer pair on even draw counts
`n_pairs = (n_draws - 1) // 2` dropped the last complete pair whenever the
draw count was even: pair m spans lags 2m and 2m+1, so the bound is
2m+1 <= n_draws-1, giving n_draws // 2 pairs. Discarding the final pair
shortens the window Geyer's initial positive sequence may examine, which
biases tau_int low and can report a spurious "no plateau" on a trajectory
that in fact resolves. Found by the test author deriving the expected pair
sums by hand rather than from a run.

Also repair three test defects surfaced by Cannon job 39384865, which was
otherwise green across the full suite (4 failed, 1814 passed):

- test_trajectory_collection built its expected tensors as float32 and
  compared them against trajectory values that ObservableTrajectory
  deliberately promotes to float64.
- the test_receipt helper collapsed a missing reason with `reason or
  <default>`, so the unresolved_without_reason and absent_without_reason
  cases silently supplied a reason and could never reach the branch they
  claimed to test. Both now use the existing _UNSET sentinel.

Refs: Task Orchestrator 123b0971-56cf-4d41-98c1-fbeafa0e0eba

+++ tests/unit/sampling/test_trajectory_collection.py
@@ -68,7 +68,9 @@ def test_collection_preserves_draw_axis_and_order() -> None:
         n_draws=3,
     )
 
-    expected = torch.tensor([[0.0, 1.0], [10.0, 11.0], [20.0, 21.0]])
+    # float64 on purpose: ObservableTrajectory promotes every trajectory so the
+    # lag sums downstream do not lose their tail to float32 rounding.
+    expected = torch.tensor([[0.0, 1.0], [10.0, 11.0], [20.0, 21.0]], dtype=torch.float64)
     assert trajectory.values.shape == (3, 2)
     torch.testing.assert_close(trajectory.values, expected)
 
@@ -87,7 +89,10 @@ def test_discard_draws_drops_only_leading_draws() -> None:
     )
 
     assert len(sampler.calls) == 4
-    torch.testing.assert_close(trajectory.values, torch.tensor([[20.0, 21.0], [30.0, 31.0]]))
+    torch.testing.assert_close(
+        trajectory.values,
+        torch.tensor([[20.0, 21.0], [30.0, 31.0]], dtype=torch.float64),
+    )
     assert trajectory.burn_in_draws == 2
